+++ config/translation.php
@@ -7,9 +7,17 @@ return [
             'resources' => '/^.+\.(?:php|vue|js)$/i',
         ]
     ],
-    'lokalise' => [
-        'apiToken' => env('LOKALISE_API_TOKEN', null),
+    'phrase' => [
+        'apiToken' => env('PHRASE_API_TOKEN', null),
         'projectId' => '',
-        'downloadOptions' => []
+        'downloadOptions' => [
+            'query' => [
+                'include_empty_translations' => true,
+                'fallback_locale_id' => 'en',
+            ]
+        ],
+        'locales' => [
+            'en',
+        ]
     ]
 ];